Resolve error about circular import
The variable `storage` that was imported from `models` package
is now being used by importing the `models` package
it is contained directly, instead of making
only the variable `storage` available
in `models/base_model.py` module.
Besides, it is said not to be a good practice to import only a variable
from a package or module.